🩹 [Patch]: Doc update and change default name on vault specific shard (#78)
## Description

This pull request includes several changes to improve the clarity of the
documentation and update configuration settings in the `Context` module.
The most important changes involve refining the descriptions in the
`README.md` file and updating the seed shard path in the configuration
file.

Documentation improvements:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L3-R28):
Enhanced the description of the `Context` concept, improved explanations
for storing module and user settings, and added more detailed usage
examples.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L3-R28)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L225-R146)

Configuration updates:

*
[`src/variables/private/Config.ps1`](diffhunk://#diff-24f682ffe9ba06c3a7c6620f5c550ddc9eac8bdfb1aefe07961c41cf1a5e9552L4-R4):
Changed the `SeedShardPath` from 'Context.shard' to 'vault.shard' for
consistency with the new naming convention.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas