🚀 [Feature]: Class for `ContextInfo` (#97)
## Description

This pull request introduces a new `ContextInfo` class and its
associated table view format to streamline the representation and
display of context-related information. The changes include defining the
class structure and adding a custom XML format for rendering the class
in a tabular view.

### New `ContextInfo` class:

*
[`src/classes/public/ContextInfo.ps1`](diffhunk://#diff-c0d9cec44494337fa03fe17fed9e3a5ebfb8232d6049aacf61042b657bfcec05R1-R15):
Added the `ContextInfo` class with properties (`ID`, `Path`, `Vault`,
`Context`) and constructors for initialization, including one that
accepts a `PSCustomObject`.

### Table view format for `ContextInfo`:

*
[`src/formats/ContextInfo.Format.ps1xml`](diffhunk://#diff-126e6b32df7ca1ad5a9c2d76448f53350aebf8ff80ae4cff329b8dca66b707f5R1-R33):
Created an XML configuration to define a table view for `ContextInfo`,
displaying `Vault` and `ID` as columns.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [x] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas